refactor and make more DRY
