+++ python/helper.py
@@ -1,0 +1,24 @@
+import re
+import json
+
+# Replacer function
+def replace_chunk(content, marker, chunk):
+    replacer = re.compile(
+        r"<!\-\- {} starts \-\->.*<!\-\- {} ends \-\->".format(marker, marker),
+        re.DOTALL,
+    )
+    chunk = "<!-- {} starts -->\n{}\n<!-- {} ends -->".format(marker, chunk, marker)
+    return replacer.sub(chunk, content)
+
+# Methods
+def ord(n):
+    return str(n)+("th" if 4<=n%100<=20 else {1:"st",2:"nd",3:"rd"}.get(n%10, "th"))
+
+
+def dtStylish(dt,f):
+    return dt.strftime(f).replace("{th}", ord(dt.day))
+
+
+def pprint(string):
+    json_formatted_str = json.dumps(string, indent=2)
+    print(json_formatted_str)
