+++ python/stocks.py
@@ -1,5 +1,5 @@
 # import
-import re
+import helper
 import random
 import json
 import pathlib
@@ -9,14 +9,6 @@ root = pathlib.Path(__file__).parent.parent.resolve()
 with open( root / "config/stocks.json", 'r') as filehandle:
   stocks_list = json.load(filehandle)
 
-# Methods
-def replace_chunk(content, marker, chunk):
-    replacer = re.compile(
-        r"<!\-\- {} starts \-\->.*<!\-\- {} ends \-\->".format(marker, marker),
-        re.DOTALL,
-    )
-    chunk = "<!-- {} starts -->\n{}\n<!-- {} ends -->".format(marker, chunk, marker)
-    return replacer.sub(chunk, content)
 
 def get_stocks(set_of_tickers):
     string_builder = ""
@@ -29,5 +21,5 @@ if __name__ == "__main__":
     index_page = root / "index.html"
     index_contents = index_page.open().read()
     string_output = get_stocks(stocks_list)
-    final_output = replace_chunk(index_contents, "stocks_marker", f"<ul>\n{string_output}</ul>")
+    final_output = helper.replace_chunk(index_contents, "stocks_marker", f"<ul>\n{string_output}</ul>")
     index_page.open("w").write(final_output)
