+++ python/weather.py
@@ -1,5 +1,5 @@
 import os
-import re
+import helper
 import pathlib
 import json
 import requests
@@ -24,18 +24,9 @@ string_today =  f"Today's date is {output_date}, Here is your daily briefing..."
 string_today += f"The average temperature today is {today_weather}˚C with highs of {high_temp}˚C and lows of {low_temp}˚C. "
 string_today += f"You can expect {today_desc} for the day."
 
-# Replacer function
-def replace_chunk(content, marker, chunk):
-    replacer = re.compile(
-        r"<!\-\- {} starts \-\->.*<!\-\- {} ends \-\->".format(marker, marker),
-        re.DOTALL,
-    )
-    chunk = "<!-- {} starts -->\n{}\n<!-- {} ends -->".format(marker, chunk, marker)
-    return replacer.sub(chunk, content)
-
 # processing
 if __name__ == "__main__":
     index_page = root / "index.html"
     index_contents = index_page.open().read()
-    final_output = replace_chunk(index_contents, "day_marker", string_today)
+    final_output = helper.replace_chunk(index_contents, "day_marker", string_today)
     index_page.open("w").write(final_output)
